+++ CHANGELOG.md
@@ -56,7 +56,7 @@
 
 ## 1.0.1 - 2018-10-05
 ### Added
-- Can now pre populate settings with a content migration
+- Can now pre-populate settings with a content migration
 
 ## 1.0.0 - 2018-10-05
 ### Added